refactor: modularize market overview
